Changing value to Float

+++ resources/resources/historical.rb
@@ -20,4 +20,4 @@ attribute :memory_kb, kind_of: Integer, default: 8388608
 attribute :rmi_address, kind_of: String, default: '127.0.0.1'
 attribute :rmi_port, kind_of: String, default: '9091'
 attribute :ipaddress, kind_of: String, default: '127.0.0.1'
-attribute :maxsize, kind_of: Integer, default: 10737418240
+attribute :maxsize, kind_of: Float, default: 10737418240.0
